feat: add sticky table of contents
